Mark order how send to integration system

+++ EasyStart/EasyStart/Logic/IntegrationSystem/SendNewOrderResult/INewOrderResult.cs
@@ -10,6 +10,6 @@ namespace EasyStart.Logic.IntegrationSystem.SendNewOrderResult
     {
         bool Success { get; }
         string ErrorMessgae { get; }
-        int OrderNumber { get; }
+        long OrderNumber { get; }
     }
 }
